refactor 2
